# frontend/contract/swagger.yaml
openapi: 3.1.0
info:
  title: Trailarr API
  summary: Trailarr API available commands.
  description: API for Trailarr application.
  termsOfService: https://github.com/nandyalu/trailarr
  contact:
    name: Trailarr - Github
    url: https://github.com/nandyalu/trailarr
  license:
    name: GNU GPL 3.0
    url: https://github.com/nandyalu/trailarr/blob/main/LICENSE
  version: 0.4.0
servers:
  - url: '{protocol}://{hostpath}'
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
      hostpath:
        default: 127.0.0.1:7889
paths:
...
  /api/v1/media/all:
    get:
      tags:
        - Media
      summary: Get All Media
      description: "Get all media from the database. \n\nOptionally apply filters\
        \ and sorting. \n\nArgs:\n    movies_only (bool, Optional=None):\n       \
        \ Flag to get only movies.\n        - If `True`, it will return only `movies`.\n\
        \        - If `False`, it will return only `series`.\n        - If `None`,\
        \ it will return all media items. \n\n    filter_by (str, Optional=`all`):\n\
        \        Filter the media items by a column value. Available filters are\n\
        \        - `all`\n        - `downloaded`\n        - `monitored`\n        -\
        \ `missing`\n        - `unmonitored`. \n\n    sort_by (str, Optional=None):\
        \ Sort the media items by `title`, `year`,             `added_at`, or `updated_at`.\
        \ \n\n    sort_asc (bool, Optional=True): Flag to sort in ascending order.\
        \ \n\nReturns:\n    list[MediaRead]: List of media objects."
      operationId: get_all_media_api_v1_media_all_get
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: movies_only
          in: query
          required: false
          schema:
            anyOf:
              - type: boolean
              - type: 'null'
            title: Movies Only
        - name: filter_by
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            default: all
            title: Filter By
        - name: sort_by
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Sort By
        - name: sort_asc
          in: query
          required: false
          schema:
            type: boolean
            default: true
            title: Sort Asc
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MediaRead'
                title: Response Get All Media Api V1 Media All Get
        '401':
          description: API key is missing
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/media/:
    get:
      tags:
        - Media
      summary: Get Recent Media
      description: "Get recent media from the database. \n\nArgs:\n    limit (int,\
        \ Optional=30): Number of items to return.\n    offset (int, Optional=0):\
        \ Number of items to skip.\n    movies_only (bool, Optional=None):\n     \
        \   Flag to get only movies.\n        - If `True`, it will return only `movies`.\n\
        \        - If `False`, it will return only `series`.\n        - If `None`,\
        \ it will return all media items. \n\nReturns:\n- list[MediaRead]: List of\
        \ media objects."
      operationId: get_recent_media_api_v1_media__get
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 30
            title: Limit
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            title: Offset
        - name: movies_only
          in: query
          required: false
          schema:
            anyOf:
              - type: boolean
              - type: 'null'
            title: Movies Only
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MediaRead'
                title: Response Get Recent Media Api V1 Media  Get
        '401':
          description: API key is missing
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/media/updated:
    get:
      tags:
        - Media
      summary: Get Updated After
      description: "Get media updated after a certain datetime. \n\nArgs:\n    seconds\
        \ (int): Seconds since epoch to filter media.\nReturns:\n    list[MediaRead]:\
        \ List of media objects."
      operationId: get_updated_after_api_v1_media_updated_get
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: seconds
          in: query
          required: true
          schema:
            type: integer
            title: Seconds
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MediaRead'
                title: Response Get Updated After Api V1 Media Updated Get
        '401':
          description: API key is missing
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/media/downloaded:
    get:
      tags:
        - Media
      summary: Get Recently Downloaded
      description: "Get recently downloaded media from the database. \n\nArgs:\n \
        \   limit (int, Optional=30): Number of items to return.\n    offset (int,\
        \ Optional=0): Number of items to skip. \n\nReturns:\n    list[MediaRead]:\
        \ List of media objects."
      operationId: get_recently_downloaded_api_v1_media_downloaded_get
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 30
            title: Limit
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            title: Offset
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MediaRead'
                title: Response Get Recently Downloaded Api V1 Media Downloaded Get
        '401':
          description: API key is missing
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/media/search:
    get:
      tags:
        - Media
        - Search
      summary: Search Media
      description: "Search media by query. \n\nArgs:\n    query (str): Search query.\
        \ \n\nReturns:\n    list[SearchMedia]: List of search media objects."
      operationId: search_media_api_v1_media_search_get
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            title: Query
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchMedia'
                title: Response Search Media Api V1 Media Search Get
        '401':
          description: API key is missing
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/media/{media_id}:
    get:
      tags:
        - Media
      summary: Get Media By Id
      description: "Get media by ID. \n\nArgs:\n    media_id (int): ID of the media\
        \ item. \n\nReturns:\n    MediaRead: Media object."
      operationId: get_media_by_id_api_v1_media__media_id__get
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: media_id
          in: path
          required: true
          schema:
            type: integer
            title: Media Id
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaRead'
        '401':
          description: API key is missing
        '404':
          description: Media Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/media/{media_id}/files:
    get:
      tags:
        - Media
      summary: Get Media Files
      description: "Get media files by ID. \n\nArgs:\n    media_id (int): ID of the\
        \ media item. \n\nReturns:\n    FolderInfo|str: Folder information or error\
        \ message."
      operationId: get_media_files_api_v1_media__media_id__files_get
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: media_id
          in: path
          required: true
          schema:
            type: integer
            title: Media Id
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/FolderInfo'
                  - type: string
                title: Response Get Media Files Api V1 Media  Media Id  Files Get
        '401':
          description: API key is missing
        '404':
          description: Media Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/media/{media_id}/download:
    post:
      tags:
        - Media
      summary: Download Media Trailer
      description: "Download trailer for media by ID. \n\nArgs:\n    media_id (int):\
        \ ID of the media item.\n    profile_id (int): ID of the trailer profile to\
        \ use for downloading.\n    yt_id (str, Optional=\"\"): YouTube ID of the\
        \ trailer.\n\nReturns:\n    str: Downloading trailer message."
      operationId: download_media_trailer_api_v1_media__media_id__download_post
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: media_id
          in: path
          required: true
          schema:
            type: integer
            title: Media Id
        - name: profile_id
          in: query
          required: true
          schema:
            type: integer
            title: Profile Id
        - name: yt_id
          in: query
          required: false
          schema:
            type: string
            default: ''
            title: Yt Id
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: string
                title: Response Download Media Trailer Api V1 Media  Media Id  Download
                  Post
        '401':
          description: API key is missing
        '404':
          description: Media Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/media/{media_id}/monitor:
    post:
      tags:
        - Media
      summary: Monitor Media
      description: "Monitor media by ID. \n\nArgs:\n    media_id (int): ID of the\
        \ media item.\n    monitor (bool, Optional=True): Monitor status. \n\nReturns:\n\
        \    str: Monitoring message."
      operationId: monitor_media_api_v1_media__media_id__monitor_post
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: media_id
          in: path
          required: true
          schema:
            type: integer
            title: Media Id
        - name: monitor
          in: query
          required: false
          schema:
            type: boolean
            default: true
            title: Monitor
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: string
                title: Response Monitor Media Api V1 Media  Media Id  Monitor Post
        '401':
          description: API key is missing
        '404':
          description: Media Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/media/{media_id}/update:
    post:
      tags:
        - Media
      summary: Update Yt Id
      description: "Update YouTube ID for media by ID. \n\nArgs:\n    media_id (int):\
        \ ID of the media item.\n    yt_id (str): YouTube ID of the trailer. \n\n\
        Returns:\n    str: Updating YouTube ID message."
      operationId: update_yt_id_api_v1_media__media_id__update_post
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: media_id
          in: path
          required: true
          schema:
            type: integer
            title: Media Id
        - name: yt_id
          in: query
          required: true
          schema:
            type: string
            title: Yt Id
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: string
                title: Response Update Yt Id Api V1 Media  Media Id  Update Post
        '401':
          description: API key is missing
        '404':
          description: Media Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '406':
          description: Invalid YouTube URL/ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ErrorResponse:
      properties:
        message:
          type: string
          title: Message
      type: object
      required:
        - message
      title: ErrorResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SearchMedia:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        year:
          type: integer
          title: Year
        youtube_trailer_id:
          type: string
          title: Youtube Trailer Id
        imdb_id:
          type: string
          title: Imdb Id
        txdb_id:
          type: string
          title: Txdb Id
        is_movie:
          type: boolean
          title: Is Movie
        poster_path:
          anyOf:
            - type: string
            - type: 'null'
          title: Poster Path
      type: object
      required:
        - id
        - title
        - year
        - youtube_trailer_id
        - imdb_id
        - txdb_id
        - is_movie
        - poster_path
      title: SearchMedia
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
...
  securitySchemes:
    APIKeyQuery:
      type: apiKey
      in: query
      name: api_key
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-KEY